Module polkadot_sdk_docs::reference_docs

Reference documents covering in-depth topics across the Polkadot SDK. It is suggested to read these on-demand, while you are going through the guides or other content.

Polkadot SDK Reference Docs.

This is the entry point for all reference documents that enhance one’s learning experience in the Polkadot SDK.

Note that this module also contains the glossary.

What is a “reference document”?

First, see why we use rust-docs for everything and our documentation principles. We acknowledge that as much of the crucial information should be embedded in the low level rust-docs. Then, high level scenarios should be covered in crate::guides. Finally, we acknowledge that there is a category of information that is:

crucial to know.

is too high level to be in the rust-doc of any one type, trait or fn.

is too low level to be encompassed in a crate::guides.

We call this class of documents “reference documents”. Our goal should be to minimize the number of “reference” docs, as they incur maintenance burden.

Modules

blockchain_state_machines

Learn about the way Substrate and FRAME view their blockchains as state machines.

chain_spec_genesis

Learn about chain specification file and the genesis state of the blockchain. Chain spec and genesis build.

cli

Learn about Substrate’s CLI, and how it can be extended.

consensus_swapping

Learn about Substrate’s consensus algorithms, and how you can switch between two. Consensus Swapping

extrinsic_encoding

Learn about how extrinsics are encoded to be transmitted to a node and stored in blocks.

fee_less_runtime

Learn about how to make a pallet/runtime that is fee-less and instead uses another mechanism to control usage and sybil attacks.

frame_benchmarking_weight

Learn about benchmarking and weight.

frame_composite_enums

Learn about composite enums in FRAME-based runtimes, such as “RuntimeEvent” and “RuntimeCall”.

frame_currency

Learn about the currency-related abstractions provided in FRAME. FRAME Currency Abstractions and Traits

frame_origin

Learn about “Origin” A topic in FRAME that enables complex account abstractions to be built.

frame_runtime_migration

Learn about all the advance ways to test your coordinate a rutnime upgrade and data migration.

frame_system_accounts

Learn about how frame-system handles account-ids, nonces, consumers and providers.

glossary

The glossary.

light_nodes

Learn about light nodes, how they function, and how Substrate-based chains come light-node-first out of the box.

metadata

Learn about metadata, the main means through which an upgradeable runtime communicates its properties to the outside world.

runtime_vs_smart_contract

Learn about the differences between smart contracts and a FRAME-based runtime. They are both “code stored onchain”, but how do they differ? Runtime vs. Smart Contracts

safe_defensive_programming

Learn about how to write safe and defensive code in your FRAME runtime.

signed_extensions

Learn about the signed extensions that form a part of extrinsics. Signed extensions are, briefly, a means for different chains to extend the “basic” extrinsic format with custom data that can be checked by the runtime.

trait_based_programming

Learn how Substrate and FRAME use traits and associated types to make modules generic in a type-safe manner.

wasm_memory

Learn about all the memory limitations of the WASM runtime when it comes to memory usage.

wasm_meta_protocol

Learn about the WASM meta-protocol of all Substrate-based chains.